refactor: 로드된 voucher 데이터를 변환하는 로직을 이넘으로 이동
